Move game math to include
