+++ src/game/model/handler/CombatHandler.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 
 #include "Handler.h"
-#include "game/model/handler/include/gamemath.h"
+#include "math/gamemath.h"
 #include "random/Random.h"
 
 constexpr Number MAXLEVEL = 100;
