+++ src/include/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(database)
 add_subdirectory(datatypes)
 add_subdirectory(graph)
+add_subdirectory(math)
 add_subdirectory(random)
 add_subdirectory(state-machine)
