+++ src/include/math/gamemath.cpp
